+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Insanely configurable logging for Node.js
 - Fix: Options documentation in JSON formatter
 - Added: Human formatter
 - Added: CommonInfoModel formatter
+- Added: Syslog formatter
 
 ### v0.1.0
 - **Initial Release**